Implement search using filter
a more concise way of implementing the search functionality. When no region is found, filter returns an empty array and (fortunately) unlike most programming languages, Javascript does not throw an exception when the index is out of bound, but returns "undefined"
Undefined is evaluated as "false" in the ternary expression. 
i,e: `
 const arr = [];
arr[0]; // returns undefined
`
Make sure to test it before merging. I couldn't replicate the testing environment in my local computer.